cleaned and commented, should be all working
