+++ config/carrotboard.yaml
@@ -1,7 +1,7 @@
-leaderboard_message_id: 894269125856215000
+leaderboard_message_id: 894868540153299000
 leaderboard_channel_id: 894267823780683900
-carrotboard_alert_channel_id: 894267846106964000
-guild_id: 1
-carrot_emoji: hello
+carrotboard_alert_channel_id: 894267846106963998
+guild_id: 884747109935497236
+carrot_emoji: 🐶
 minimum_carrot_count: 1
 minimum_pin_count: 1
