+++ events/cb_reactionRemove.js
@@ -11,10 +11,16 @@ module.exports = {
      * @param {User} user
      */
     async execute(reaction, user) {
+        // check if partial
+        if (reaction.partial) {
+            reaction = await reaction.fetch();
+        }
+
         /** @type {CarrotboardStorage} */
         const cbStorage = global.cbStorage;
         const message = reaction.message;
         
+        // make sure not bot and not the current client
         if (!message.author.bot && !reaction.me) {
             // get the details
             const emoji = reaction.emoji.toString();
