+++ lib/carrotboard/index.js
@@ -17,7 +17,7 @@ class CarrotboardStorage {
     /** @param {Client} client */
     constructor(client) {
         this.config = new CarrotboardConfig();
-        this.db = new DBcarrotboard();
+        this.db = {}// new DBcarrotboard();
         this._client = client;
     }
 
@@ -27,6 +27,7 @@ class CarrotboardStorage {
      * @param {String} emoji
      */
     async sendCBAlert(messageReaction, CBID, emoji) {
+        // get the alert channel and check it
         let alertChannel;
         try {
             alertChannel = await this._client.channels.fetch(this.config.alertChannelID);
@@ -40,9 +41,6 @@ class CarrotboardStorage {
             console.error("Carrotboard: sendCBAlert: Alert channel is not a Guild text channel.");
             return;
         }
-        
-        /** @type {TextChannel} */ 
-        alertChannel;
 
         // get the message content
         const message = messageReaction.message;
@@ -79,6 +77,7 @@ class CarrotboardStorage {
     }
 
     async updateLeaderboard() {
+        // generate the leaderboard
         const embedPages = await this.generateLeaderboard({ onlyFirstPage: true, emoji: this.config.carrot });
         const embed = embedPages[0];
 
@@ -99,6 +98,7 @@ class CarrotboardStorage {
             return;
         }
 
+        // update the leaderboard
         await message.edit({embeds: [embed]});
     }
 
@@ -168,21 +168,20 @@ class CarrotboardStorage {
             }
 
             // add the row to the embed
-            embedPages[pageNum]
-            .addFields([
+            embedPages[pageNum].addFields([
                 {
                     name: `${entryIndex}: ${messageAuthor.username}`,
-                    value: `CB ID: ${cbID}`,
+                    value: `[ID: ${cbID}](${url})\n` + "\u2800".repeat(6),
                     inline: true
                 }, 
                 {
-                    name: "Message",
-                    value: `${content}\n[Goto](${url})`,
+                    name: "Message" + "\u2800".repeat(6), // 20 for full but messed up on thin display 
+                    value: `${content}\n\u200b`,
                     inline: true
                 },
                 {
                     name: `Number of ${emoji}`,
-                    value: count,
+                    value: `${count}\n\n\n\u200b`,
                     inline: true
                 }
             ]);
@@ -190,6 +189,7 @@ class CarrotboardStorage {
             entryIndex += 1;
         }
 
+        // there werent any results, add the empty embed
         if (embedPages.length == 0) {
             const sadEmbed = new MessageEmbed({title: "There are no Carroted Messages :( :sob: :smiling_face_with_tear:", description: " "});
             embedPages.push(sadEmbed);
@@ -242,6 +242,7 @@ class CarrotboardConfig {
     _pinMinimum;
 
     constructor() {
+        // read the file
         const file = fs.readFileSync("./config/carrotboard.yaml", "utf8");
         const parsed = YAML.parse(file, { intAsBigInt: true });
         const expectedKeys = {
@@ -254,6 +255,7 @@ class CarrotboardConfig {
             "minimum_pin_count": "pinMinimum",
         };
 
+        // check the config keys
         for (const key in expectedKeys) {
             const value = parsed[key];
             if (value == null) {
@@ -264,6 +266,7 @@ class CarrotboardConfig {
     }
 
     saveToFile() {
+        // convert into yaml
         const yamlStr = YAML.stringify({
             "leaderboard_message_id": BigInt(this.leaderboardID),
             "leaderboard_channel_id": BigInt(this.permaChannelID),
@@ -274,6 +277,7 @@ class CarrotboardConfig {
             "minimum_pin_count": Number(this.pinMinimum)
         });
 
+        // write it to the file
         fs.writeFile("./config/carrotboard.yaml", yamlStr, function(err) {
             if (err) {
                 console.error(err);
@@ -350,24 +354,29 @@ class CarrotboardConfig {
     get pinMinimum() { return this._pinMinimum }
 }
 
-/** @param {String} messageStr */
+/** 
+ * Returns the first emoji seen in the message, or null sometimes
+ * @param {String} messageStr
+ */
 function extractOneEmoji(messageStr) {
     if (messageStr.length == 0) {
         return null;
     }
 
+    // the emoji regexs
     const customEmojiRegex = /<a?:\w+:\d{18}>/;
     const normalEmojiRegex = /(?:[\u2700-\u27bf]|(?:\ud83c[\udde6-\uddff]){2}|[\ud800-\udbff][\udc00-\udfff])[\ufe0e\ufe0f]?(?:[\u0300-\u036f\ufe20-\ufe23\u20d0-\u20f0]|\ud83c[\udffb-\udfff])?(?:\u200d(?:[^\ud800-\udfff]|(?:\ud83c[\udde6-\uddff]){2}|[\ud800-\udbff][\udc00-\udfff])[\ufe0e\ufe0f]?(?:[\u0300-\u036f\ufe20-\ufe23\u20d0-\u20f0]|\ud83c[\udffb-\udfff])?)*/;
     // `normalEmojiRegex` yoinked from "https://medium.com/reactnative/emojis-in-javascript-f693d0eb79fb"
     
+    // match the regexs
     const customMatch = messageStr.match(customEmojiRegex);
     const normalMatch = messageStr.match(normalEmojiRegex);
-    
     if (customMatch != null && normalMatch != null) {
         // both occured, bad
         return null;
     }
 
+    // get the winning match
     const match = customMatch ?? normalMatch;
     if (match == null) {
         // none occured, bad
